add InventorySceneRenderer
